+++ NEWS.md
@@ -1,6 +1,7 @@
 # h5lite 2.1.0.0
 
-* Added support for SZIP compression - see `vignette('data-compression')`.
+* New functions: `h5_compression()` and `h5_inspect()`.
+* Added support for a variety of compression codecs - see `vignette('compression')`.
 * Added support for partial reading - see `vignette('partial-reading')`.
 * `h5_length()` and `h5_dim()` now return numeric values instead of integer.